+++ src/main/java/org/chocosolver/util/iterators/EvtScheduler.java
@@ -17,11 +17,6 @@ import org.chocosolver.solver.variables.events.IEventType;
 public interface EvtScheduler<E extends IEventType> extends IntIterator {
 
     /**
-     * Initializes the scheduler with an event defined by 'type'.
-     * @param type event received
-     */
-    void init(E type);
-    /**
      * Initializes the scheduler with an event defined by 'mask'.
      * @param mask event's mask received
      */