+++ src/test/java/org/chocosolver/solver/constraints/nary/sum/IntLinCombTest.java
@@ -18,7 +18,6 @@ import org.chocosolver.solver.constraints.Operator;
 import org.chocosolver.solver.constraints.Propagator;
 import org.chocosolver.solver.constraints.nary.cnf.PropTrue;
 import org.chocosolver.solver.exception.ContradictionException;
-import org.chocosolver.solver.propagation.PropagationEngineFactory;
 import org.chocosolver.solver.search.strategy.Search;
 import org.chocosolver.solver.variables.BoolVar;
 import org.chocosolver.solver.variables.IntVar;
@@ -280,7 +279,6 @@ public class IntLinCombTest {
     @Test(groups="1s", timeOut=60000, dataProvider = "boolean")
     public void testUSum2(boolean incr) throws ContradictionException {
         Model sum = sum(new int[][]{{-2, 7}, {-1, 6}, {2}, {-2, 5}, {-2, 4}, {-2, 6}}, new int[]{-7, 13, -3, -18, -24, 1}, 30, 0, incr);
-        PropagationEngineFactory.DEFAULT.make(sum);
         Variable[] vars = sum.getVars();
         ((IntVar) vars[0]).instantiateTo(-2, Cause.Null);
         ((IntVar) vars[1]).instantiateTo(-1, Cause.Null);